feat: UserService 추상화, 트랜잭션 적용 구현체 추가
